+++ The-Ultimate-React-Native-Series-Fundamentals/4-Styling/done-with-it/App.js
@@ -1,23 +1,7 @@
-import React from 'react';
-import { View, StyleSheet } from 'react-native';
-import Card from './app/components/Card';
+import ListingDetailsScreen from "./app/screens/ListingDetailsScreen";
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Card
-        title="Red jacket for sale"
-        subtitle="$100"
-        image={require('./app/assets/jacket.jpg')}
-      />
-    </View>
+    <ListingDetailsScreen></ListingDetailsScreen>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#f8f4f4',
-    padding: 20,
-    paddingTop: 100,
-  },
-});
