--- The-Ultimate-React-Native-Series-Fundamentals/4-Styling/done-with-it/app/components/AppText/AppText.js
@@ -1,9 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-import styles from "./styles";
-
-function AppText({ children }) {
-  return <Text style={styles.text}>{children}</Text>;
-}
-
-export default AppText;
