--- The-Ultimate-React-Native-Series-Fundamentals/4-Styling/done-with-it/app/components/AppText/index.js
@@ -1,3 +1,0 @@
-import AppText from "./AppText";
-
-export default AppText;
