--- The-Ultimate-React-Native-Series-Fundamentals/4-Styling/done-with-it/app/components/AppText/style.js
@@ -1,19 +1,0 @@
-import { StyleSheet, Platform } from "react-native";
-
-const styles = StyleSheet.create({
-  text: {
-    color: "tomato",
-    ...Platform.select({
-      ios: {
-        fontSize: 20,
-        fontFamily: "Avenir",
-      },
-      android: {
-        fontSize: 18,
-        fontFamily: "Roboto",
-      },
-    }),
-  },
-});
-
-export default styles;
