+++ The-Ultimate-React-Native-Series-Fundamentals/4-Styling/done-with-it/app/screens/ListingDetailsScreen.js
@@ -1,0 +1,47 @@
+import React from 'react';
+import { View, Image, StyleSheet } from 'react-native';
+import AppText from '../components/AppText';
+import colors from '../config/colors';
+
+function ListingDetailsScreen() {
+  return (
+    <View>
+      <Image
+        source={require('../assets/jacket.jpg')}
+        style={styles.image}
+      />
+
+      <View style={styles.detailsContainer}>
+        <AppText style={styles.title}>
+          Red jacket for sale
+        </AppText>
+
+        <AppText style={styles.price}>
+          $100
+        </AppText>
+      </View>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  image: {
+    width: '100%',
+    height: 300,
+  },
+  detailsContainer: {
+    padding: 20,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: '500',
+  },
+  price: {
+    color: colors.secondary,
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginVertical: 10,
+  },
+});
+
+export default ListingDetailsScreen;
